refactor(tools): fold annotation data into TOOL_MANIFEST (#3597)
Increment 2 of #3563. Move the per-tool { annotations, sideEffects } records
from tool-annotations.ts INTO TOOL_MANIFEST entries (each is now
{ name, annotations, sideEffects }), and derive TOOL_ANNOTATIONS +
getToolAnnotations/getMcpAnnotations from the manifest. The annotation TYPES move
to the manifest (the import-free leaf); tool-annotations.ts re-exports them for
existing consumers. Single source of truth; closes #3563 down to done.

Value preservation: the new manifest body was GENERATED programmatically from the
current data (no hand-transcription), and the derived TOOL_ANNOTATIONS was
deep-compared against a pre-refactor golden master — ZERO value mismatches across
all 46 tools. Reference identity preserved for the #3358 live accessor.

Blast radius (all updated + verified):
- AST parser scripts/parse-tool-manifest.ts now extracts name from object-literal
  elements (+ a fixture test); REGISTERED_TOOL_NAMES and gap-detector
  AVAILABLE_TOOLS derive via .map(t => t.name); RegisteredToolName = [number]['name'].
- inject-governance.ts: checkToolAnnotations + checkToolPrerequisites + the
  non-read-only extractor read the manifest; the two meta-tests mutate the manifest.
- generate-repo-index.ts + generate-docs-content.ts adopt the shared AST parser
  (their line-regex would pick up titles/descriptions); generate-tool-reference.ts
  + check-mcp-description-drift.ts iterate .name.

byte-identical: governance:check 46 tools, docs:tools:check 47 files, repo-index +
docs-content + description-drift all clean. 244 script tests + 2194 mcp/tools tests
pass; typecheck + lint clean.

Closes #3597.
