+++ skills/documentation-management/SKILL.md
@@ -79,6 +79,7 @@ allowed-tools: Read, Edit, Write, Bash, Grep, Glob, Task
 <!-- PIPELINE NOTE: link-check.yml gained a Rule-of-Two secret-gating guard (#3778, 2026-06-08) — its `link-check` job now skips pull_request runs from `auto-remediation/*` head refs (the auto-remediation plan doc is markdown and matches the job's path filter, and the job carries GITHUB_TOKEN). No change to push/schedule/dispatch link validation. Mirrors the existing ci.yml/pr-review.yml guards; kept in sync with AUTO_REMEDIATION_BRANCH_PREFIX. A regression test (auto-remediation-branch.test.ts) asserts all four secret-bearing PR-triggered workflows carry the guard. -->
 <!-- PIPELINE NOTE: inject-governance.ts extractMcpTools() switched from a line-oriented regex to an AST walk (#3596, 2026-06-09) — the new pure module scripts/parse-tool-manifest.ts parses the canonical TOOL_MANIFEST array (then REGISTERED_TOOL_NAMES, then a legacy `tools:` property) via the TypeScript compiler API, returning string-literal elements in source order. Formatting-agnostic (comments between elements, single-line arrays, quote style) and the seam that lets the list become a fully *derived* value later (a regex over a literal cannot read a derived value — the #3566 5/5 vote). Output is byte-identical: governance:check unchanged at 46 tools. Parser carries 9 fixture unit tests (scripts/parse-tool-manifest.test.ts). -->
 <!-- PIPELINE NOTE: docs-check.yml gained a "Tool Reference Drift" job (#3689, 2026-06-09) that runs `scripts/generate-tool-reference.ts --check` (= `pnpm docs:tools:check`) and fails when the committed docs/reference/tools/ output is stale. Closes the doc-drift loop for the #3687 generated per-tool MCP reference, mirroring the repo-index/governance --check gates (regenerate in memory, fail on diff). The push trigger's paths filter now also lists scripts/generate-tool-reference.ts. Regenerate + commit via `pnpm docs:tools`. -->
+<!-- PIPELINE NOTE: TOOL_MANIFEST folded annotation/side-effect data in (#3597, 2026-06-09) — entries are now `{ name, annotations, sideEffects }` objects, not bare name strings, and TOOL_ANNOTATIONS derives from the manifest. The pipeline scripts that read tool names from the manifest all moved to the shared AST parser scripts/parse-tool-manifest.ts (which now also extracts `name` from object-literal elements): inject-governance.ts (checkToolAnnotations now reads the manifest, not tool-annotations.ts), generate-repo-index.ts and generate-docs-content.ts (replaced their line-regex with parseRegisteredToolNames), generate-tool-reference.ts + check-mcp-description-drift.ts (iterate `.name`). All counts byte-identical (governance:check 46 tools, docs:tools:check 47 files). Adding/removing a tool = edit the manifest array. -->
 
 **Full specification:** [docops-spec.md](../../docs/ops/docops-spec.md)
